Update patch set 7
Patch Set 7:

We tested this on UNH Ampere POD 1. We're not sure this is all it takes for the installation to begin due to some network issues with that POD (which we hope to solve after Thanksgiving), but it's the first set of changes required for the rootfs to be dumped to disk correctly.
Jimmy also tested this locally in his lab and got further along the deployment process, up to the point the hw-detector and ironic-virtmedia-driver changes are needed to support the Ampere hardware (patches will be sent for review soon for those too).

Patch-set: 7
Label: Verified=0